+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_json_writer_factory_nlohmann.h
@@ -1,0 +1,26 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <memory>
+
+#include "opentelemetry/exporters/otlp/otlp_json_writer.h"
+#include "opentelemetry/exporters/otlp/otlp_json_writer_factory.h"
+#include "opentelemetry/version.h"
+
+OPENTELEMETRY_BEGIN_NAMESPACE
+namespace exporter
+{
+namespace otlp
+{
+
+class OPENTELEMETRY_EXPORT JsonWriterFactoryNlohmann : public JsonWriterFactory
+{
+public:
+  std::unique_ptr<JsonWriter> Create() override;
+};
+
+}  // namespace otlp
+}  // namespace exporter
+OPENTELEMETRY_END_NAMESPACE